refactor: add aliases support
